Problem: pulp_rpm_prerequisites is slow to install packages
Solution: Use yum module for RH family, the only supported family
currently, and fallback to package for others.

Also remove epel-release: We now rely on ansible-pulp to install
it beforehand due to how this role is now included, and since
packages are no longer being installed 1 by 1, it will not help.

[noissue]